# openapi/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.ImportProjectRequest:
    properties:
      projects:
        items:
          $ref: '#/definitions/models.Project'
        type: array
    type: object
...
  models.NotificationResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      message:
        type: string
      notification_type:
        type: string
      read:
        type: boolean
      read_at:
        type: string
      title:
        type: string
    type: object
...
  models.Project:
    properties:
      address:
        type: string
      commission_rate:
        type: number
      created_at:
        type: string
      delivery_date:
        type: string
      description:
        type: string
      guid:
        type: string
      id:
        type: integer
      interest_rate:
        type: number
      lot_count:
        type: integer
      lots:
        description: Associations
        items:
          $ref: '#/definitions/models.Lot'
        type: array
      measurement_unit:
        type: string
      name:
        type: string
      price_per_square_unit:
        type: number
      project_type:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Fintera Real Estate Management System
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fintera API
  version: "1.0"
paths:
...
  /audits:
    get:
      consumes:
      - application/json
      description: Get a paginated list of system audit logs
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 50
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Audit Logs
      tags:
      - Audit
...
  /notifications:
    get:
      consumes:
      - application/json
      description: Get a paginated list of notifications for the current user
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Notifications
      tags:
      - Notifications
  /notifications/{notification_id}:
    delete:
      consumes:
      - application/json
      description: Delete a notification
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Notification
      tags:
      - Notifications
    get:
      consumes:
      - application/json
      description: Get a notification by ID
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.NotificationResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Notification
      tags:
      - Notifications
    put:
      consumes:
      - application/json
      description: Mark a notification as read
      parameters:
      - description: Notification ID
        in: path
        name: notification_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark Notification Read
      tags:
      - Notifications
  /notifications/mark_all_as_read:
    post:
      consumes:
      - application/json
      description: Mark all notifications as read for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Mark All Notifications Read
      tags:
      - Notifications
...
  /projects:
    post:
      consumes:
      - application/json
      description: Create a new project
      parameters:
      - description: Project Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.ProjectResponse'
      security:
      - BearerAuth: []
      summary: Create Project
      tags:
      - Projects
  /projects/{project_id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Project
      tags:
      - Projects
    get:
      consumes:
      - application/json
      description: Get a project by ID
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Project
      tags:
      - Projects
    put:
      consumes:
      - application/json
      description: Update an existing project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Project Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ProjectResponse'
      security:
      - BearerAuth: []
      summary: Update Project
      tags:
      - Projects
  /projects/{project_id}/approve:
    post:
      consumes:
      - application/json
      description: Approve a project (Admin)
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Approve Project
      tags:
      - Projects
  /projects/{project_id}/lots:
    get:
      consumes:
      - application/json
      description: Get a paginated list of lots for a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: per_page
        type: integer
      - description: Search term
        in: query
        name: search_term
        type: string
      - description: Filter by status
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List Lots
      tags:
      - Lots
    post:
      consumes:
      - application/json
      description: Create a new lot in a project
      parameters:
      - description: Project ID
        in: path
        name: project_id
        required: true
        type: integer
      - description: Lot Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Lot'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.LotResponse'
      security:
      - BearerAuth: []
      summary: Create Lot
      tags:
      - Lots
  /projects/{project_id}/lots/{lot_id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a lot
      parameters:
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Lot
      tags:
      - Lots
    get:
      consumes:
      - application/json
      description: Get a lot by ID
      parameters:
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LotResponse'
      security:
      - BearerAuth: []
      summary: Get Lot
      tags:
      - Lots
    put:
      consumes:
      - application/json
      description: Update an existing lot
      parameters:
      - description: Lot ID
        in: path
        name: lot_id
        required: true
        type: integer
      - description: Lot Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.Lot'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.LotResponse'
      security:
      - BearerAuth: []
      summary: Update Lot
      tags:
      - Lots
...
  /projects/import:
    post:
      consumes:
      - application/json
      description: Bulk import projects
      parameters:
      - description: Projects Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.ImportProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Import Projects
      tags:
      - Projects
  /reports/commissions_csv:
    get:
      description: Download commissions report as CSV
      produces:
      - text/csv
      responses:
        "200":
          description: commissions.csv
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Commissions Report
      tags:
      - Reports
  /reports/customer_record_pdf:
    get:
      description: Download customer record PDF (Hoja de Cliente)
      parameters:
      - description: Contract ID
        in: query
        name: contract_id
        required: true
        type: integer
      produces:
      - application/pdf
      responses:
        "200":
          description: customer_record.pdf
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Customer Record PDF
      tags:
      - Reports
  /reports/overdue_payments_csv:
    get:
      description: Download overdue payments report as CSV
      produces:
      - text/csv
      responses:
        "200":
          description: overdue.csv
          schema:
            type: file
      security:
      - BearerAuth: []
      summary: Overdue Payments Report
      tags:
      - Reports
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"